+++ Runtime/AthlosWebView.cs
@@ -4,10 +4,21 @@ namespace Athlos
 {
   public abstract class AthlosWebView : MonoBehaviour
   {
+    public enum Target
+    {
+      None,
+      Event,
+      MatchSeries,
+      User
+    }
+    
     [Header("Initial Colors")]
     [SerializeField] private bool setInitialColors;
     [SerializeField] private Color backgroundColor;
     [SerializeField] private Color spinnerColor;
+      
+    private Target _target;
+    private string _targetId;
 
     private string BaseURL
     {
@@ -18,6 +29,29 @@ namespace Athlos
 #else
         return "https://inapp.athlos.gg";
 #endif
+      }
+    }
+    
+    private string TargetURLDir
+    {
+      get
+      {
+        switch (_target)
+        {
+          case Target.Event:
+          {
+            return $"/events/{_targetId}";
+          }
+          case Target.MatchSeries:
+          {
+            return $"/match/{_targetId}";
+          }
+          case Target.User:
+          {
+            return $"/users/{_targetId}";
+          }
+        }
+        return "";
       }
     }
     
@@ -38,7 +72,7 @@ namespace Athlos
     {
       get
       {
-        string url = $"{BaseURL}?tenant={Athlos.Tenant}{InitialColorParams}";
+        string url = $"{BaseURL}{TargetURLDir}?tenant={Athlos.Tenant}{InitialColorParams}";
 #if UNITY_EDITOR || DEVELOPMENT_BUILD
         if (Athlos.TestAuthentication)
         {
@@ -62,5 +96,19 @@ window.__athlos_auth = () => {{
 }}";
       }
     }
+    
+    /// <summary>
+    /// Open Athlos at a specific item such as an event, match or player
+    /// </summary>
+    /// <param name="target">The type of the item to open Athlos at</param>
+    /// <param name="id">The unique ID of the item to open</param>
+    public void OpenAt(Target target, string id)
+    {
+      _target = target;
+      _targetId = id;
+      OnInitialURLChanged();
+    }
+    
+    protected abstract void OnInitialURLChanged();
   }
 }